implemented new background design
liked the idea of the game being played on a wooden table, so that's that, and I also decided to use colors instead of an image because it makes more sense in terms of screen size, but it can always look prettier, so any criticisms? Will keep implementing new stuff (like color palette, new container design) later.